Update patch set 4
Patch Set 4:

Hi Nancy,

You may want to keep an eye on nightly test to see if there is timeout after this CL get submitted as it enables tests that were disabled. I guess the speed up for realtime tests would make up for the tests on 444 clips.

Thanks!

Patch-set: 4
CC: Gerrit User 8673 <8673@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>