rtc: Updates to cyclic refresh
Move the postencode update stats to the end of
encode_recode_loop, and add metric for amount of
low motion in frame and sequence.
Adjust the GF period and the disabling of refresh
based on the new metric.

For speed 6/7/8: Neutral on average metrics for
rtc/rtc_derf set, but gains on the high motion clip
(street): ~2% on rtc, ~4% on rtc_derf.

Change-Id: I3fc4ae67bfc527bb9a26f4b25d99b7ac43283ec4