Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Urvang,
> 
> We are seeing negligible change in scaling with unaligned loads for aom_sad_16x32_sse2.
> 
>   Module name	                Previous scaling        Current scaling 
> 			            w.r.t. C               w.r.t. C
>                       	
> aom_sad_16x32_sse2	              1.11	              1.10
> (windows)
> aom_sad_16x32_sse2                   26.87	             26.55
> (linux)
> 
> Note: For linux we have observed that the scaling of C vs SSE2 is huge as compared to windows. The reason might be that gcc compiler in case of linux is not optimizing the C module that well.

Correction:

Hi Urvang,

We are seeing negligible change in scaling with unaligned loads for aom_sad_16x32_sse2.

  Module name	                Previous scaling        Current scaling 
			            w.r.t. C               w.r.t. C
                      	
aom_sad_16x32_sse2	              1.11	              1.10
(windows)
aom_sad_16x32_sse2                   55.22	             52.23
(linux)
Note: For linux we have observed that the scaling of C vs SSE2 is huge as compared to windows. The reason might be that gcc compiler in case of linux is not optimizing the C module that well.

Patch-set: 1