Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Ruiling, thanks for the patch. I ran a quick test before/after the patch, and gathered instruction count result.
> 
> perf stat -e instructions:u ./aomenc -o out.webm BasketballDrill_832x480_50.y4m --end-usage=vbr --target-bitrate=800 --cpu-used=4 --threads=0 --profile=0 --lag-in-frames=19 --min-q=0 --max-q=63 --auto-alt-ref=1 --passes=2 --kf-max-dist=150 --kf-min-dist=0 --drop-frame=0 --static-thresh=0 --bias-pct=50 --minsection-pct=0 --maxsection-pct=2000 --arnr-maxframes=7 --arnr-strength=5 --sharpness=0 --undershoot-pct=100 --overshoot-pct=100 --frame-parallel=0 --tile-columns=0 --limit=60
> before: 341,793,130,940
> after:  342,160,462,769
> 
> perf stat -e instructions:u ./aomenc -o out.webm night_720p30.y4m --end-usage=vbr --target-bitrate=2500 --cpu-used=4 --threads=0 --profile=0 --lag-in-frames=19 --min-q=0 --max-q=63 --auto-alt-ref=1 --passes=2 --kf-max-dist=150 --kf-min-dist=0 --drop-frame=0 --static-thresh=0 --bias-pct=50 --minsection-pct=0 --maxsection-pct=2000 --arnr-maxframes=7 --arnr-strength=5 --sharpness=0 --undershoot-pct=100 --overshoot-pct=100 --frame-parallel=0 --tile-columns=0 --limit=60
> before: 1,189,135,412,846
> after:  1,190,158,738,452
> 
> Looks like it increases a little. Could you double check? Will share the clips with you.
> p.s. speed 4 & 5 was combined recently, so new speed 4 is the old speed 5.

ok, I will try to check that, it may need to count the instructions generated by compiler. But do run a performance test to see how much total encoding time used on your side may help. I guess simply counting instructions may not work here, instructions operate on registers and instructions operate on memory have quite different performance.
What compiler do you use?

Let me explain why the patch works. The bottleneck in previous implementation is the three for-loops. all the a/b/c are allocated on the stack. so you have to read-write the stack memory for each shuffle operation. And even if the loop-unrolling in compiler help to unroll the loop, the compiler still could not hold all the data in register(there are totally only 16 ymm registers). I see gcc failed to unroll the loops, I guess unrolling may help a lot here. But still should not be as good as the new implementation.
The new implementation load all the data into registers(only 8 ymm registers needed to hold the data) and do the shuffle work in registers. all the loops in the new implementation are easily unrolled by gcc. So the reason why the new version may be better is mainly because all the instructions operate on registers while the previous version operate data in stack memory.

Patch-set: 1