Use is_mi_coded map to determine has_bottom_left in intra coding
For AV2 baseline, no result changed should be expected.
When 3WAY_PARTITIONS tool is enabled, a slight performance change
is introduced because of a corner case: for the 2nd block of a
horizontal 3way partition, for example, 16x8 in a 3-way partitioned
16x16, if 4x4 tx is applied, the new algorithm will determine that
the bottom-left tx block does have its bottom-left reference
available if the left column is available.

Change-Id: Ib6f5f7d3183a2b3f5a17abb1be166dc1a091df17