Write aom_image ITU-T T.35 metadata to stream
Adds av1_write_metadata_array() to write metadata payload corresponding
to an aom_image. It will be written to the stream just before the frame
OBU that has the show_frame set. Only ITU-T T.35 case is implemented.

Adds test to check if metadata obu was written correctly to bitstream.

BUG=aomedia:2507

Change-Id: I94ae89fc63c30b2d86ab23e9689b97b55e367592