Update patch set 1
Patch Set 1:

Hello Tom,
This is the proposal for image metadata writing. It's based on the needs for HDR10+ metadata but the place to write it could be adapted for a more general case with your suggestion.

Patch-set: 1
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>