Change metadata type from uint8_t to uint32_t.
The metadata_type syntax element in the bitstream is of the type leb128(),
which can be from 0 to 1^32 - 1. It would be good to declare the
metadata type field or function parameter as uint32_t to be
future-proof.

Also declare the "uint8_t *data" parameter of aom_img_add_metadata() and
aom_img_metadata_alloc() as const.

BUG=aomedia:2507

Change-Id: If00379639119788da04bfc9cd56a3980314db742