Update patch set 19
Patch Set 19:

> Patch Set 18:
> 
> Hi Victoria, thanks for the optimization. This CL is huge. My suggestion is to separate it into small CLs(the smaller, the better) to make it easy for reviewing. Also, if something is wrong later, we don't have to revert the whole thing.
> Also, could you please provide some speed test result if 1 CL is meant to cause speed change?
Yunqing, thanks a lot!
This commit is just 1/3 of the whole thing actually - I do understand you concern and did the split prior to committing :) But OK, will do 1 more split here. As for performance increase - it is there but not so impressive as the whole one coming that's why I haven't named it... But right you are again - the cat in the black box couldn't be accepted, I will do my best to measure something.Stay tuned!

Patch-set: 19