Update patch set 6
Patch Set 6:

(1 comment)

> Patch Set 5: Code-Review+2
> 
> Does the change make max_delta larger or smaller in general?

I think that depends where you are in the clip.

Note though that looking at it I do not think max_delta can ever be greater in magnitude that vbr_bits_off_target with this change so the following clause can be simplified. Also I think a check is needed for frame_window > 0.

Patch-set: 6