Update patch set 6
Patch Set 6:

Apologies - only the aarch64 specializations are implicated. Per comment #24 on the bug:
https://bugs.chromium.org/p/aomedia/issues/detail?id=2261#c24
For Jenkins multi-target job, we have temporary put in a filter to mask off all specific arm64 code for function 'av1_convolve_2d_sr_neon' in file ./av1/common/arm/convolve_neon.c. This was just done for this unit test to see if the problem persist. This was done on July 15th. So far we have not encounter 
the any crash on arm64.

The check looks like this:
# remove aarch64 ifdef for intermittent unit test failure - 
# ( ImageBrightnessTests/EdgeDetectBrightnessTest.BlurUniformBrightness)
sed -i -e '847,1407s;defined(__aarch64__);0  // &;' ${WORKSPACE}/aom/av1/common/arm/convolve_neon.c
    

For the given range it changes any #if defined(__aarch64__) to #if 0

Patch-set: 6