Update patch set 7
Patch Set 7:

> Code under __arm64__ is computing for multiple rows at a time.
> We will review and see if there are any assumptions w.r.t. height being multiple of 4 or 8.
> Once this is merged and hack in jenkins is removed, we can bring back these optimizations after required changes and asserts() and see if the problem is seen again.

Thanks. That works for me.

When working on the convolve for libgav1 I noticed that vertical pass (in 2 pass) does not use a multiple of 4:
https://chromium.googlesource.com/codecs/libgav1/+/refs/heads/master/src/dsp/arm/convolve_neon.cc#916
https://chromium.googlesource.com/codecs/libgav1/+/refs/heads/master/src/dsp/arm/convolve_neon.cc#962
Notably we need to calculate (up to) 7 extra rows to be used as context in the vertical pass.

There we used vext instead of a transpose to allow doing one row at a time:
https://chromium.googlesource.com/codecs/libgav1/+/refs/heads/master/src/dsp/arm/convolve_neon.cc#229

Patch-set: 7