Remove twopass->kfgroup_inter_fraction
I believe that the intended function of this factor was to take account
of the overall level of predictability in estimating the active max q.
This assumes that for better predicted content the key frame will have
a higher boost and will consume a larger proportion of the total bits for
the KF group. In fact this number does not vary that much for most longer
KF groups and is dominated instead by the length of the key frame group.

In this patch I remove the factor and work in a value that is similar to the
average seen when coding 150 frame groups in out tests.

This patch also separates the calculation of bits per MB used for estimating
maxq from that used in the normal per frame rate control as the correction
mechanisms are substantially different.

Metrics wise there is a mixed picture with NET gains in overall psnr and ssim
but average psnr and psnr-hvs are down a bit.

The exception is the HD set where there are gains on all metrics. This is
because 3 of the clips have a small KF group at the start and see gains of
5-12%, which is enough to distort the averages for the whole set. This
also illustrates how kfgroup_inter_fraction can go wrong.

Quite large changes to the unit test thresholds expected because the unit
tests have very short kf groups.

STATS_CHANGED

Change-Id: I71391f2a0d06ba9ce82f13a8f09366297bec14fc