Remove twopass->kfgroup_inter_fraction
I believe that the intended function of this factor was to take account
of the overall level of predictability in estimating the active max q.
This assumes that for better predicted content the key frame will have
a higher boost and will consume a larger proportion of the total bits for
the KF group. In fact this number does not vary that much for most longer
KF groups and is dominated instead by the length of the key frame group.
This causes some odd effects in some clips.

In this patch I remove the factor and work in a value that is similar to the
average seen when coding 150 frame groups in out tests.

This patch also separates the calculation of bits per MB used for estimating
maxq from that used in the normal per frame rate control as the correction
mechanisms are substantially different.

STATS_CHANGED

Change-Id: I71391f2a0d06ba9ce82f13a8f09366297bec14fc