rtc: Fixes to rate control for CBR mode
For CBR mode: constrain the active_best for gf/arf
refresh frames on the gf_cbr_boost_pct flag, as this
logic was intended only for boosted gf/arfframes.
Also move the QP adjustmemnt to separate function and
fix the conditions for the adjustment/clamping.

Neutral change in bdrate for speed >= 6 for rtc.

Reduces some excessive overshoot in some cases,
also some reduction in pulsing artifact.

Change-Id: I981c452752c7bde61a25c24dfa52d4cfaae48ed1