Rate control: a fusion of Q mode and VBR mode
For the VBR mode, we determine q for each frame using the Q mode's
mechanism, with cq_level set to the active_worst_quality determined
at each gop start.

Idea behind this change:
We observe that Q mode's compression performance is better than the
VBR mode. There's potential room for VBR mode.

Therefore, I think we can simply borrow the mechanism that the Q mode
determines q for each frame. We feed in the active_worst_quality
estimated by the VBR mode as the cq_level.

We still rely on the feedback loop of the VBR mode to control bit
rate target.

Manual inspection suggests that this change doesn't break the rate
control. The accuracy stays similar as compared to the baseline.

spd4        Avg_psnr    ovr_psnr    ssim      vmaf
Lowres      -0.11       -0.08       0.20      -0.44
Midres       0.03        0.02       0.59      -0.01
Hdres       -0.03        0.04       0.14       0.09
Ugc360p      0.08       -0.03       0.36      -0.05

There is not noticable gain after the change of undershoot and
overshoot range from 25 to 100 for the Ugc360p dataset.

This CL is the minimum change to make the fusion of the Q mode and
the VBR mode work. If this CL is approved, we need more refactoring
and cleaning CLs to ensure the change is compatible with other modes
and paths as well, for example, CBR mode.

STATS_CHANGED

Change-Id: I40dc21309a1e3f94fda961c363c0e305f7849a66