rtc: Fix to avg_frame_low_motion in cyclic refresh
The avg_frame_low_motion should be updated every
frame if the AQ_MODE=CR feature is enabled.
In current code the metric was not updated when
the cyclic refresh was dynamically disabled
within the stream. This can lead to the refresh
not being re-enabled when motion content goes low.

Also adjust the threshold for disabling CR.

Only affects a few high motion clips on rtc set.
For low bitrates speed 7: 3.5% gain for pixel3 clip,
0.46% gain for still_bright, and 0.78% loss for street.

Change-Id: I0e33861c98f29e109c2e755932f0b8ac87598bc8