Update patch set 5
Patch Set 5:

(3 comments)

> Patch Set 5:
> 
> (1 comment)

Thanks for the comments!
I was CC'ed you added a patch to monorail, from your comments the idea seems the same, but I don't have permission to access the issue page. I had a proposal but hadn't pushed it because I'm not familiar with oss-fuzz or Gerrit process and was unsure if it needed to be a new commit with new change ID, a revert of the revert or send a patch as you did in monorail... so I'll just wait for your proposal and comment on it if it's OK since it see it covers the issues. If some action is needed from my side let me know.

Patch-set: 5