Make last intra-inter mode explicit.
The last mode in the intra-inter-modes is sometimes used to determine
whether the predictor needs to be recomputed (since buffers are re-used).
Instead of hard-coding it to II_SMOOTH_PRED, use INTERINTRA_MODES - 1.
There is no functional change, but the experimental branch is expanding
the number of INTERINTRA_MODES, and this change will help keep the two
in sync.

Change-Id: Ib43de4f0b38430bf48d12bcec388b1db9f69bcf2