Adjust rate control enumerator.
This adjustment is to improve behavior when stricter rate
control is specified on the command line.

When the undershoot and overshoot limits are set to 25% the
metrics improvements are as follows.

(Av psnr, Ov psnr, SSIM Psnr-hvs)
Low Res  -0.947	-1.032	-0.490	-1.092
UGC 360  -0.587	-0.735	0.229	-0.966
Mid Res  -0.335	-0.408	-0.127	-0.517
Hd Res   -1.012	-1.108	-0.224	-0.930

However the setting best for 25% os/us is a little worse for 100%
which was the setting previously used to tune this number.

Low Res  0.068	0.149	0.166	0.047
UGC 360  0.357	0.226	0.936	-0.035
Mid Res  -0.061	0.101	0.248	-0.293
Hd Res   0.025	0.236	0.060	0.019

So why the difference  ?

In general using tighter rate control hurts metrics. If we are targeting
the number of bits spent on the clip closely we may spend extra bits,
for example, on frames where there is little objective or subjective benefit
simply because of undershoot on earlier frames.

Two rate control metrics come into play here. One (1) is forward looking and if
there was undershoot or overshoot in previous ARF groups it tries to do better
in subsequent groups, but does not try and re-allocate bits. The second (2) more
specifically targets the total bits spent in the clip by reallocation of the bit budget.

Lowering the value of the enumerator used in choosing the active Q range a little
seems to improve rate accuracy from (1) slightly and this has a small negative impact
on metrics when the rate error tolerance is 100 (which effectively disables (2)).

However, when we set a tolerance of 25% the mechanism for (2) is activated less with
a lower enumerator and the improvements in metrics that result more than offset the
negative impact. Hence the observations here reflect the balance between the two rate
control methods and using (1) more aggressively and (2) less seems to give a better
outcome.

Note also that the reduction of the enumerator compared that used in maked sense in
regard to the expectation that AV1 has improved compression efficiency.

Still Being tuned so do not submit yet.

STATS_CHANGED

Change-Id: I9c9cfcfce66e74b739aac370c97c6391128b30bb