Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> It makes sense that lowering the enumerator could improve the accuracy when there's undershooting. What about the overshooting? Is it getting worse?

Yes you would think thought the adaptation range of rc_factor (which I still want to replace and thin is not working that well) is greater for overshoot for some reason in the code at the moment.

On the low res set the clips that saw an improvement in rate accuracy with the 25% setting, as well as a metrics improvement outnumbered those with a drop in rate accuracy about 2:1. For the UGC360P set where there were previously more examples of overshoot than undershoot there were 27 where the rate accuracy dropped and 17 where it improved by my counting but again with metrics improvement.

I plan to follow up with other patches to try and refine the behavior.

Patch-set: 2