Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > Hi, Yunqing, I do remember about changes, I will rebase, however all my questions - about unit tests and separate files requirement don't depend on a particular version - this one or a rebased one.
> 
> Victoria, it is requested that an unit test is submitted in the same simd optimization patch. This is a recent optimization patch. Maybe you can use it as a reference.
> https://aomedia-review.googlesource.com/c/aom/+/100643

Yes, I do understand unit tests policy- it is code quality standard, and that's why I'm asking. It is not difficult, but it is very time consuming :( and it may last forever if I start doing it,  so the question is - is there any chances to ask some (Google) intern doing that?

Patch-set: 7