Change to calculation of rc_factor.
This change has a modest negative impact on metrics both when
the undershoot and overshoot % is set to 100 and when it is set to
25 (0.1 to 0.4%). However, it substantially improves rate accuracy
for the vast majority clips in all our test sets.

The trade off of improved rate accuracy vs metrics impact seems
to be good and there are metrics gains if the os/us % is
adjusted so that pre and post patch have more comparable rate
accuracy.

More detailed metrics results to follow.

STATS_CHANGED

Change-Id: Id50c8855a403dff6ee1585fe9c148050d549b1ac