rtc: Fix to get_ref_frame_flags for nonrd mode
For nonrd mode where ALT_REF is used (speed 7),
we can remove the GOLDEN reference if its the same
as LAST or ALT_REF.

This gives a small speedup as we avoid checking
golden ref in nonrd_pickmode on some frames (where
golden coincides with alt_ref).

Neutral change in bdrate.
Speedup ??

Change-Id: I6f0e23363eff57d3c3c2069734fbda5f7f287550