Update patch set 4
Patch Set 4:

> Patch Set 4: -Code-Review
> 
> I'm trying to understand the change of RC_FACTOR_MAX from 1.75 to 1.25.
> It allows a smaller rc_factor and thus lower active_worst_q when overshooting.
> It means when there's an overshooting, we apply less strict correction to the rc_factor.
> 
> If we see better rate accuracy, does it imply that previously some clips with undershooting actually has at least a gop overshot and then too much correction is applied on the rc_factor and cause undershooting in subsequent frames?

The maximum I reduced simply to make the variability in line with the undershoot case (+/- 25%). In fact 175 is not hugely worse. The main difference (and reason we overall see more overshoot) is that the factor is applied much more consistently throughout the clip vs before. Previously the correction naturally declined as we progress through the clip. This is because the rf was calculated based on the ratio of the bit error to the total bits spent so far. Hence we get a big initial adjustment to say an undershoot but even if we continue to undershoot on a group by group basis the RF may actually move in the wrong direction because the total bits so far is growing faster than the rate error. This change means that the correction may get damped towards the middle of the clip but will grow again if need be towards the end, even if the actual bit error is declining.  This means RF tends to have more impact. It may be better to look at the ratio of the error to a fixed rate (e.g. 1 seconds bits) but so far things that logically make more sense to me have proved worse. Perhaps this is down to damping and that it really does matter more to correct aggressively at the start and if need be at the end. 

I don't think this is the end of the road by any means just something that in terms of rate accuracy vs metrics impact was better.

Patch-set: 4