Update patch set 1
Patch Set 1: Code-Review+2

So we're doing this because we're going to be adding avif tags to libaom, and the tags may have a different format than those we've been using for libaom thus far. I have some reservations about this plan, but I understand the reasoning. Thad said, I would prefer that we structure the avif tags so that this change isn't necessary. 

LGTM/+2 because there's nothing functionally wrong here and the change is correct given the example I've seen for the structure of the avif tag(s).

Patch-set: 1
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+2