Alter the worst allowed Q at different arf level.
This ensures that even if we are clamped at the active maxq
we do not end up with the same Q for low level arf frames and
leaf (non arf) frames. This is important for the TPL model
which assumes that at each level in the pyramid structure there
is a drop in Q.

There is the potential for clips that already overshoot to do so
more with this change, and for clips that currently undershoot to
do so less. In practice this should be mitigated by adjustment of
the active Q range and as shown below the net impact on rate
accuracy across tested sets was neutral.

Metrics and rate control outcomes for low res and ugc360.
(other sets pending)

(os/us 100%)
-0.035	-0.142	0.023	-0.11
-0.238	-0.361	-0.21	-0.416

(os/us 25%)
-0.035	-0.142	0.023	-0.11
-0.408	-0.42	-0.42	-0.464

Do not submit yet pending more results and adjustment.

STATS_CHANGED

Change-Id: I70ed8acc61404480e6bf49c6c31e7b4e5b40e908