Change rdopt to use zero-offset NEARMV
This change should make the encoder treat ref_mv_idx as zero-offset when
it handles NEARMV.  It is a prerequisite for allowing NEARMV with a drl
index of zero and eventually removing NEARESTMV.

This change on its own increases (is bad for compression) BD-rate by 0.5%,
likely because NEARMV signalling now takes an extra bit.  This will
hopefully not matter once NEARESTMV is removed.

Change-Id: I0a1afc1c5084d354660d29186777d2f2bd4a5849