Change rdopt to use zero-offset NEARMV.
This change should make the encoder treat ref_mv_idx as zero-offset when
it handles NEARMV.  It is a prerequisite for allowing NEARMV with a drl
index of zero and eventually removing NEARESTMV.

Currently this change is broken.
It segfaults on 17 frames of lowres/ice_cif when compiled without debug flags.

Change-Id: I0a1afc1c5084d354660d29186777d2f2bd4a5849