Add option for better interintra wedge search
The interintra mode and wedge pattern search was done separately.
This patch added an option of joint search which brings slight coding
gains with some extra encoding complexity cost. It is enabled for speed
0 and 1 only.

Coding gains:
          lowres     midres    ugc360p
speed 0    0.08%     0.07%      0.02%
speed 1    0.10%     0.15%      0.06%

Average speed overhead: 2% for speed 0; 3% for speed 1.

STATS_CHANGED for speed 0 and 1.

Change-Id: I181a4cdbf42f63bfd13fefe3b4bbcb4bd9326eca