Update patch set 6
Patch Set 6:

Hi Yunqing,

We found a bug in the existing compound_type_rd function while reusing the stored stats which is related to the calculation of modelled rd.
In the case of data reuse, modelled rd should be recalculated using the current rate_mv instead of using the previous record.
This has in turn improved the data reuse when COMPOUND_DISTWTD is enabled, and has given encode time reduction in cpu-used=1 preset.

Though there is a slight drop in bd-rate, we are submitting this patch since it is a bug fix. This will also enable us to experiment
with the reuse of compound type decision based on MV difference.

The following are the links of AWCY runs for the current patch.

Speed 1: (0.72 encode time reduction and bd-rate drop is 0.0142)
https://beta.arewecompressedyet.com/?job=dec18_ref_cpu1_76cd26%402019-12-23T01%3A55%3A30.646Z&job=dec18_bugfix_cpu1%402019-12-24T10%3A02%3A01.431Z

Speed 2: (0.10 encode time reduction and bd-rate drop is 0.0191)
https://beta.arewecompressedyet.com/?job=dec18_ref_cpu2_76cd26%402019-12-23T01%3A55%3A42.070Z&job=dec18_bugfix_cpu2%402019-12-24T10%3A01%3A51.551Z

Speed 3: (No speed impact and bd-rate drop is 0.0107)
https://beta.arewecompressedyet.com/?job=dec18_ref_cpu3_76cd26%402019-12-23T01%3A54%3A39.690Z&job=dec18_bugfix_cpu3%402019-12-24T10%3A01%3A42.303Z

Speed 4: (No speed impact and bd-rate drop is 0.0106)
https://beta.arewecompressedyet.com/?job=dec18_ref_cpu4_76cd26%402019-12-23T01%3A55%3A18.904Z&job=dec18_bugfix_cpu4%402019-12-24T10%3A01%3A17.207Z

Please note that even with this bug fix, the reuse of stats will not produce identical results as reuse disabled (link below for AWCY) since the handling of 
prune_wedge_pred_diff_based speed feature is different for NEWMV and non-NEWMV cases.
https://beta.arewecompressedyet.com/?job=dec18_ret_zero_cpu3_76cd26%402019-12-23T13%3A46%3A15.061Z&job=dec18_reuse_cpu3_76cd26%402019-12-23T01%3A56%3A00.518Z

Please review and merge the patch if no issues.

Thanks & Regards,
Venkata.

Patch-set: 6
Reviewer: Gerrit User 5185 <5185@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 9039 <9039@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 6245 <6245@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 13265 <13265@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC-email: vikas.prasad@ittiam.com