Add dropout optimization.
Dropout optimization can be viewed as a fast mode of trellis
optimization, yet not as accurate. The rational is to forcibly set some
quantized coefficients, which are in the middle of zeros, to zeros. The
main goal is to reduct the bits used for encoding but barely affect the
decoded image. Compared to trellis optimization, this is much faster.

There are two usage of dropout optimization:
(1) A further optimization after trellis optimization.
(2) An alternative to trellis optimization, especially in real-time
mode when trellis optimization is disabled due to high time complexity.

Experimental results:

Under Speed-4 (two-pass mode):
          avg PSNR   overall PSNR     SSIM
lowres       0.008         -0.009    0.029
midres      -0.005         -0.010   -0.000
ugc360p     -0.006         -0.010    0.005

Under Speed-6 (real-time mode):
          avg PSNR   overall PSNR     SSIM
lowres      -0.033         -0.029    0.058
midres      -0.072         -0.135   -0.037
ugc360p     -0.163         -0.218   -0.137

STATS_CHANGED

Change-Id: Icabd73546fefe0d0fd9dc2d112754e84d8c5d550