Make DRL cost for NEARMV 0 identical to NEARESTMV
Under this change, bitstream losses are entirely recovered under the
full 150-frame lowres test set (BD-rate change ~-0.1%).
Interestingly/shockingly/annoyingly, this change is NOT reflected under
17-frame lowres, which shows a BD-rate change of ~+0.2%.

Remaining work for this project is seeing if any further gains can be
had with more context tuning, retuning entropy tables, and refactoring
out the last vestiges of NEARESTMV.

It's also worth seeing if other patchsets have similar performance when
tested on full-frame lowres.

Otherwise, this project is now "done" in the sense that NEARESTMV is gone
and all losses are recovered.

Change-Id: I55feca6e14ef46dcf5b44e1149288274d541274b