Update patch set 13
Patch Set 13:

(2 comments)

> Patch Set 12:
> 
> (2 comments)
> 
> Good strategy and speedups. Maybe goes a little further: for a block with large source var, but skip=1(skip tx), it might be okey with no loop restoration.
> If the speedup at speed 4 is good, I will re-test speed 5 option(maybe we can enable loop restoration there).

Yunqing,

Regarding the skipping of loop restoration in case of tx skip, for a 256x256 or 128x128 LRU, the tx skip flags of all the transform blocks within the LRU need to be tracked. Please let us know if this needs to be pursued so that we can explore its scope.

Regards,
Nithya

Patch-set: 13