Update patch set 4
Patch Set 4:

Hi Max,

This appears to be another piece of application level metadata that should be handled opaquely by the library. From looking at the implementation in this patch it seems like the only thing special about HDR metadata is that it's being inserted only within temporal units that contain a sequence header (key frames basically).

Instead of adding API hooks to set data that the library doesn't really need, I suggest extending the existing metadata implementation by adding flags that instruct the library when to insert metadata. Something like this:

enum MetadataInsertionFlags {
  kNextFrame,
  kNextKeyFrame,
  // Maybe not necessary, but for example:
  kAllFrames
}

The methods for working with metadata would need to be updated to support the flags, but I think it would be a cleaner setup that ultimately allows for much more flexibility. 

The code itself is not unreasonable, but I think nearly all of it should be part of aomenc instead of being built into the library. The application using the library should be responsible for writing (or reading) well formed HDR10 metadata, and not the library itself.

Wan-Teh: WDYT?

Patch-set: 4
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>