Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > Hi Max,
> > 
> > This appears to be another piece of application level metadata that should be handled opaquely by the library. From looking at the implementation in this patch it seems like the only thing special about HDR metadata is that it's being inserted only within temporal units that contain a sequence header (key frames basically).
> > 
> > Instead of adding API hooks to set data that the library doesn't really need, I suggest extending the existing metadata implementation by adding flags that instruct the library when to insert metadata. Something like this:
> > 
> > enum MetadataInsertionFlags {
> >   kNextFrame,
> >   kNextKeyFrame,
> >   // Maybe not necessary, but for example:
> >   kAllFrames
> > }
> > 
> > The methods for working with metadata would need to be updated to support the flags, but I think it would be a cleaner setup that ultimately allows for much more flexibility. 
> > 
> > The code itself is not unreasonable, but I think nearly all of it should be part of aomenc instead of being built into the library. The application using the library should be responsible for writing (or reading) well formed HDR10 metadata, and not the library itself.
> > 
> > Wan-Teh: WDYT?
> 
> Hi Tom,
> 
> Thanks for the prompt reply and implementation suggestions.
> 
> I think it mostly comes down to your last statement and the question if libaom should be responsible for writing well formed HDR10 metadata or not.
> I created the CL giving libaom the responsibility, but your suggestion about the flexibility issue also seems valid. I can work of the change if you consider it, I was just concerned of the possible misuse and incorrect HDR10 stream generation. 

The primary concern motivating my suggested implementation changes is with avoidance of further growth of the API sprawl in libaom wherever possible. Since metadata can be handled opaquely by the library I think that's the best way for it to be handled. With it appearing that we need only to add a notion of where metadata belongs in the output stream to the existing API to achieve the desired goal of this patch I would like to go that route. 

That said I don't see any harm in moving the validation code you have for the CLL and MDCV values into aomenc. Preventing the libaom tool(s) from writing garbage into the stream is certainly worthwhile. However, beyond that, I think we must rely on users of the library to implement their own support for generation of the CLL and MDCV payloads correctly. After all, even if the support was internal to the library and it performed validation, that still would not prevent an encoder application that's using the library from ignoring the built in support and inserting its own malformed CLL/MDCV/whatever metadata OBUs into the stream. 

> 
> 
> PD. for Wan-Teh: Sorry for the misspell in my first comment, auto-correct tricked me.

Patch-set: 4
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>