Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (1 comment)
> 
> Sarah,
> total_stats is always allocated in the case of two pass. The assert is not required.

To clarify further, this assert was added by us in the patch https://aomedia-review.googlesource.com/c/aom/+/103401 , on the assumption that total_stats will not be used in the case of single pass. Since this assumption does not hold in the case of VBR single pass, we have removed the assert. Please let me know if you still feel the assert is required.

Patch-set: 3