1-pass: correctly infer length of last GF group
Earlier, 1-pass was assuming that upto 'rc->frames_to_key' are still
available in define_gf_group_pass0() call. This sometimes led to a
baseline_gf_interval larger than actual frames available for the last GF
group.

For example, even if 15 frames are available, last GF group of length 16
will be created, with ALTREF pointing beyond the end of the video. And
because of that, later on ALTREF was implicitly disabled and all frames
were coded in order.

Instead, we now use lookahead buffer, when available, to better infer
the end of the video, and thereby the length of last GF group.

STATS_CHANGED

Change-Id: I1ae574e1f7c9bb75e3baaf2cab544b4a7dd9250e